Move from a literate config to a exploded config
I'm very hesitant to do this, since I enjoyed being able to tie my various bits of config
together with overarching explanations. Ultimately, the reason why I chose to go this
route was because the editing experience for individual files was poor in
`org-mode`. While `org-edit-special` works as well as can be expected, it doesn't drop
into the full file, which means we only edit with context from the current block. This
doesn't scale well.

As I've worked on my config, it feels increasingly like comments should go in the code,
more then vise versa. Less indirection is easier for me.